requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.21.0
opencv-python>=4.5.0
scipy>=1.7.0
ml_collections
tensorboardX
tqdm
medpy
SimpleITK
Pillow
FastGeodis
matplotlib
einops
